oc-3c3: split replay history out of google runner
