--- src/agents/pi-embedded-runner.applygoogleturnorderingfix.test.ts
@@ -1,62 +1,0 @@
-import type { AgentMessage } from "@mariozechner/pi-agent-core";
-import { SessionManager } from "@mariozechner/pi-coding-agent";
-import { describe, expect, it, vi } from "vitest";
-import { applyGoogleTurnOrderingFix } from "./pi-embedded-runner.js";
-import { castAgentMessage } from "./test-helpers/agent-message-fixtures.js";
-
-describe("applyGoogleTurnOrderingFix", () => {
-  const makeAssistantFirst = (): AgentMessage[] => [
-    castAgentMessage({
-      role: "assistant",
-      content: [{ type: "toolCall", id: "call_1", name: "exec", arguments: {} }],
-    }),
-  ];
-
-  it("prepends a bootstrap once and records a marker for Google models", () => {
-    const sessionManager = SessionManager.inMemory();
-    const warn = vi.fn();
-    const input = makeAssistantFirst();
-    const first = applyGoogleTurnOrderingFix({
-      messages: input,
-      modelApi: "google-generative-ai",
-      sessionManager,
-      sessionId: "session:1",
-      warn,
-    });
-    expect(first.messages[0]?.role).toBe("user");
-    expect(first.messages[1]?.role).toBe("assistant");
-    expect(warn).toHaveBeenCalledTimes(1);
-    expect(
-      sessionManager
-        .getEntries()
-        .some(
-          (entry) =>
-            entry.type === "custom" && entry.customType === "google-turn-ordering-bootstrap",
-        ),
-    ).toBe(true);
-
-    applyGoogleTurnOrderingFix({
-      messages: input,
-      modelApi: "google-generative-ai",
-      sessionManager,
-      sessionId: "session:1",
-      warn,
-    });
-    expect(warn).toHaveBeenCalledTimes(1);
-  });
-
-  it("skips non-Google models", () => {
-    const sessionManager = SessionManager.inMemory();
-    const warn = vi.fn();
-    const input = makeAssistantFirst();
-    const result = applyGoogleTurnOrderingFix({
-      messages: input,
-      modelApi: "openai",
-      sessionManager,
-      sessionId: "session:2",
-      warn,
-    });
-    expect(result.messages).toBe(input);
-    expect(warn).not.toHaveBeenCalled();
-  });
-});
