+++ src/agents/pi-embedded-runner.openai-tool-id-preservation.test.ts
@@ -4,7 +4,7 @@ import {
   makeInMemorySessionManager,
   makeModelSnapshotEntry,
 } from "./pi-embedded-runner.sanitize-session-history.test-harness.js";
-import { sanitizeSessionHistory } from "./pi-embedded-runner/google.js";
+import { sanitizeSessionHistory } from "./pi-embedded-runner/replay-history.js";
 import { castAgentMessage } from "./test-helpers/agent-message-fixtures.js";
 
 describe("sanitizeSessionHistory openai tool id preservation", () => {
