+++ src/agents/pi-embedded-runner/sanitize-session-history.tool-result-details.test.ts
@@ -3,7 +3,7 @@ import type { ToolResultMessage, UserMessage } from "@mariozechner/pi-ai";
 import { SessionManager } from "@mariozechner/pi-coding-agent";
 import { describe, expect, it } from "vitest";
 import { makeAgentAssistantMessage } from "../test-helpers/agent-message-fixtures.js";
-import { sanitizeSessionHistory } from "./google.js";
+import { sanitizeSessionHistory } from "./replay-history.js";
 
 describe("sanitizeSessionHistory toolResult details stripping", () => {
   it("strips toolResult.details so untrusted payloads are not fed back to the model", async () => {
